+++ tests/DoctrineMigrations/Version20180101000000.php
@@ -10,15 +10,15 @@ class Version20180101000000 extends AbstractMigration
     public function up(Schema $schema)
     {
         // this up() migration is auto-generated, please modify it to your needs
-        $this->abortIf($this->connection->getDatabasePlatform()->getName() != "sqlite", "Migration can only be executed safely on 'sqlite'.");
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() != "mysql", "Migration can only be executed safely on 'mysql'.");
         
-        $this->addSql("CREATE TABLE dummy (id INTEGER PRIMARY KEY, name TEXT NOT NULL, createdAt DATETIME NOT NULL)");
+        $this->addSql("CREATE TABLE dummy (id int NOT NULL AUTO_INCREMENT, name VARCHAR (255) NOT NULL, createdAt DATETIME NOT NULL, PRIMARY KEY (id))");
     }
 
     public function down(Schema $schema)
     {
         // this down() migration is auto-generated, please modify it to your needs
-        $this->abortIf($this->connection->getDatabasePlatform()->getName() != "sqlite", "Migration can only be executed safely on 'sqlite'.");
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() != "mysql", "Migration can only be executed safely on 'mysql'.");
         
         $this->addSql("DROP TABLE dummy");
     }