+++ tests/exportedSql/expected_configuration.sql
@@ -1,0 +1,4 @@
+SET FOREIGN_KEY_CHECKS=0;
+INSERT INTO `table_1` VALUES ('test','me'),('test','you');
+
+SET FOREIGN_KEY_CHECKS=1;